Rename namespaces, clean up using directives
