+++ LibDescent/Data/FixVector.cs
@@ -20,7 +20,7 @@
     SOFTWARE.
 */
 
-namespace PiggyDump
+namespace LibDescent.Data
 {
     public struct FixAngles
     {
