+++ LibDescent/Data/HXMFile.cs
@@ -22,11 +22,9 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.IO;
 
-namespace PiggyDump
+namespace LibDescent.Data
 {
     public struct ReplacedBitmapElement
     {
