+++ LibDescent/Data/PIGImage.cs
@@ -21,15 +21,11 @@
 */
 
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Drawing;
 using System.Drawing.Imaging;
 using System.IO;
-using System.Windows.Forms;
 
-namespace PiggyDump
+namespace LibDescent.Data
 {
     public class PIGImage
     {
