+++ LibDescent/Data/POFReader.cs
@@ -20,13 +20,9 @@
     SOFTWARE.
 */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.IO;
 
-namespace PiggyDump
+namespace LibDescent.Data
 {
     public class POFReader
     {
