+++ LibDescent/Data/Palette.cs
@@ -20,16 +20,10 @@
     SOFTWARE.
 */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.IO;
-using System.Windows.Forms;
 using System.Drawing;
 using System.Drawing.Imaging;
 
-namespace PiggyDump
+namespace LibDescent.Data
 {
     public class Palette
     {
