+++ LibDescent/Data/RLEEncoder.cs
@@ -20,14 +20,7 @@
     SOFTWARE.
 */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Drawing;
-using System.Windows.Forms;
-
-namespace PiggyDump
+namespace LibDescent.Data
 {
     public class RLEEncoder
     {
