+++ LibDescent/Data/SNDFile.cs
@@ -22,12 +22,9 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.IO;
 
-namespace PiggyDump
+namespace LibDescent.Data
 {
     public struct SoundData
     {
